% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mlr3_vars.R
\name{add_mlr3_vars}
\alias{add_mlr3_vars}
\title{Add mlr3 variables to an \code{rmlx_splits} object}
\usage{
add_mlr3_vars(splits)
}
\arguments{
\item{splits}{(rmlx_splits)}
}
\value{
(rmlx_splits)
}
\description{
Add mlr3 variables to an \code{rmlx_splits} object
}
