% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_constraints.R
\name{combine_constraints}
\alias{combine_constraints}
\alias{cc}
\title{Combine multiple constraints}
\usage{
combine_constraints(..., constraints = list(), type = NULL, target = NULL)

cc(..., constraints = list(), type = NULL, target = NULL)
}
\arguments{
\item{...}{(character | rmlx_constraint | rmlx_constraints) ...}

\item{constraints}{(list) \cr further arguments not passed via \code{...} (with same allowed classes).}

\item{type}{(NULL | character) \cr ...}

\item{target}{(NULL | character) \cr ...}
}
\value{
(rmlx_constraints)
}
\description{
...
}
\examples{
combine_constraints(
  specify_constraint("test$age > 50"),
  specify_constraint("test$risk == 'high'")
)
}
