% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_metrics.R
\name{derive_metrics}
\alias{derive_metrics}
\title{Derive metrics for all data splits and models.}
\usage{
derive_metrics(predictions, splits, ..., by = c("idx_split", "idx_model"))
}
\arguments{
\item{predictions}{(data.frame) data with variables defined by \code{by} and predictions and labels.
For instance the result of \code{\link{get_mlr3_predictions}}.}

\item{splits}{(rmlx_splits) splits object}

\item{...}{(...) calculation of metrics based on variables in \code{predictions},
will be passed to \code{dplyr::reframe}}

\item{by}{(character) grouping variables defining the associated predictions}
}
\value{
(data.frame), usually passed on \code{\link{get_mlr3_predictions}}.
}
\description{
Derive metrics for all data splits and models.
}
