% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_splits.R
\name{derive_splits}
\alias{derive_splits}
\title{Derive a data split for a given estimand.}
\usage{
derive_splits(design, data)
}
\arguments{
\item{design}{(rmlx_design) \cr a \code{rmlx_design} object, i.e. created via
\code{\link{specify_design}()}. Alternatively, directly pass an \code{rmlx_estimand},
\code{rmlx_splitting_rule} or \code{rmlx_spec_nested} which will be converted into an \code{rmlx_design}.}

\item{data}{(data.frame) \cr the complete data set.}
}
\value{
(rmlx_splits)
}
\description{
Derive a data split for a given estimand.
}
