% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_splits.R
\name{describe_splits}
\alias{describe_splits}
\title{Describe the derived data splits}
\usage{
describe_splits(splits, data, ...)
}
\arguments{
\item{splits}{(rmlx_splits) \cr obtained via \code{\link{derive_splits}}}

\item{data}{(data.frame) \cr data that was used to derive the \code{splits}}

\item{...}{(quote) \cr named \code{\link{quote}}s that describe properties of training or test sets.}
}
\value{
data.table, an extended version of get_info(splits)
}
\description{
Describe the derived data splits
}
