% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate synthetic data for testing purposes}
\usage{
generate_data(n_obs = 2000)
}
\arguments{
\item{n_obs}{(numeric) \cr number of observations}
}
\value{
(data.frame)
}
\description{
Generate synthetic data for testing purposes
}
\examples{
generate_data(200)
}
