% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_.R
\name{get_info}
\alias{get_info}
\title{Get an overview of basic splits information}
\usage{
get_info(splits)
}
\arguments{
\item{splits}{(rmlx_splits) \cr result from \code{\link{derive_splits}}.}
}
\value{
data.table with information
}
\description{
Get an overview of basic splits information
}
