% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mlr3_vars.R
\name{get_mlr3_}
\alias{get_mlr3_}
\alias{get_mlr3_sets_train}
\alias{get_mlr3_sets_test}
\alias{get_mlr3_sets}
\alias{get_mlr3_map}
\alias{get_mlr3_vars}
\title{Retrieve objects needed for mlr3 training/tuning.}
\usage{
get_mlr3_sets_train(splits)

get_mlr3_sets_test(splits)

get_mlr3_sets(splits)

get_mlr3_map(splits)

get_mlr3_vars(splits)
}
\arguments{
\item{splits}{(rmlx_splits) \cr the splits object}
}
\value{
Depends on the function name:
\itemize{
\item sets_train: list indices of train sets.
\item sets_test: list indices of test sets.
\item sets = list(train=sets_train, test=sets_test)
\item map: data.table which gives the idx_split for each test observation
\item vars: list(sets=sets, map=map)
}
}
\description{
Retrieve objects needed for mlr3 training/tuning.
}
\details{
To avoid re-calculation of these objects, consider adding them to the splits object via
\code{\link{add_mlr3_vars}}.
}
