% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{get_valid_idx}
\alias{get_valid_idx}
\title{Get valid ids}
\usage{
get_valid_idx(splits)
}
\arguments{
\item{splits}{(rmlx_splits) \cr an object created with \code{\link{derive_splits}}}
}
\value{
(numeric) \cr a vector with valid indices
}
\description{
Get valid ids
}
