% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_nested.R
\name{nest}
\alias{nest}
\title{specify a nested data splitting scheme}
\usage{
nest(outer, inner = outer, name = get_default_name(outer, inner))
}
\arguments{
\item{outer}{(rmlx_spec_simple) \cr outer specification}

\item{inner}{(rmlx_spec_simple) \cr inner specification}

\item{name}{(character) \cr name of the nested specification}
}
\value{
(rmlx_spec_nested)
}
\description{
specify a nested data splitting scheme
}
