% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_constraint.R
\name{parse_constraints}
\alias{parse_constraints}
\title{Parse constraints}
\usage{
parse_constraints(x)
}
\arguments{
\item{x}{(character) \cr a character expression, can be augmented by further "#" parameters, see examples}
}
\value{
(rmlx_constraints)
}
\description{
...
}
\examples{
parse_constraints("test$age >= 18 # name=ic_age #type=population")
}
