% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{\%,\%}
\alias{\%,\%}
\title{Pipe operator for concatenation}
\usage{
lhs \%,\% rhs
}
\arguments{
\item{lhs}{(any) \cr an R object}

\item{rhs}{(any) \cr an R object such that \code{c(lhs, rhs)} is defined, usually of same class as \code{lhs}.}
}
\value{
The result of calling \code{c(lhs, rhs)}.
}
\description{
Pipe operator for concatenation
}
