% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain.R
\name{preproc_expr}
\alias{preproc_expr}
\title{Preprocess expr argument in specify_constraint()}
\usage{
preproc_expr(expr, target = c("test", "relation", "train"))
}
\arguments{
\item{expr}{(character | formula) expression to be checked}

\item{target}{(character) \cr either "test", "relation" or "train"}
}
\value{
(character) \cr the preprocessed expression
}
\description{
Preprocess expr argument in specify_constraint()
}
