% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_models.R
\name{select_models}
\alias{select_models}
\title{Select models based on the results of the inner loop of a nested data splitting scheme.}
\usage{
select_models(
  data,
  metric,
  aggregate = mean,
  select = max,
  groups = NULL,
  break_ties = "random",
  args_aggregate = list(),
  args_select = list()
)
}
\arguments{
\item{data}{(data.frame) Result from \code{derive_metrics} or similar data.frame.}

\item{metric}{(character)}

\item{aggregate}{(function) How should results be aggregated across inner folds? (Default: mean)}

\item{select}{(function) Which model should be chosen? (e.g. \code{max} for performance metrics,
\code{min} for error metrics)}

\item{groups}{(character) \if{html}{\out{<experimental>}} should the selection be conducted per group?}

\item{break_ties}{(character) How should ties be broken? (Default: "randomly")}

\item{args_aggregate}{(list) further named arguments passed to \code{aggregate}}

\item{args_select}{(list) further named arguments passed to \code{select}}
}
\value{
(data.frame) Benchmark results for selected models
}
\description{
Select models based on the results of the inner loop of a nested data splitting scheme.
}
