% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain.R
\name{specify_constraint}
\alias{specify_constraint}
\alias{sc}
\title{Specify constraints on data sets or observations}
\usage{
specify_constraint(
  expr,
  name = to_character(expr),
  vars = detect_vars(expr),
  target = detect_target(expr),
  type = c(NA, "context", "population", "admissibility")
)

sc(
  expr,
  name = to_character(expr),
  vars = detect_vars(expr),
  target = detect_target(expr),
  type = c(NA, "context", "population", "admissibility")
)
}
\arguments{
\item{expr}{(character | formula) \cr ...}

\item{name}{(character) \cr ...}

\item{vars}{(character) \cr ...}

\item{target}{(character) \cr ...}

\item{type}{(character) \cr ...}
}
\value{
(rmlx_constraint) \cr ...
}
\description{
A short description...
}
\examples{
specify_constraint("test$age > 18", type = "population")
specify_constraint(~ (test$year - train$year) \%>\% dplyr::between(1, 3), type = "context")
specify_constraint("nrow(train) > 100")
}
