% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_nested.R
\name{specify_design}
\alias{specify_design}
\title{Specify a simple or nested experimental design}
\usage{
specify_design(spec)
}
\arguments{
\item{spec}{(rmlx_spec) \cr specify a simple (non-nested) design by providing
either an \code{rmlx_estimand} or \code{rmlx_splitting_rule}. To specify a nested design,
use \code{nest()}.}
}
\value{
(rmlx_design)
}
\description{
Specify a simple or nested experimental design
}
\examples{
e <- specify_estimand(specify_constraint(~ train$country != test$country))
s <- specify_splitting_rule("cv", n_folds = 5)

# simple, structured (estimand-based) design:
specify_design(e)

# simple, unstructured (splitting-based) design:
specify_design(s)

# nested, structured design (nested CV in this case):
specify_design(nest(s))

# nested, hybrid design:
specify_design(nest(e, s))

}
