% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_estimand.R
\name{specify_estimand}
\alias{specify_estimand}
\title{Specify an Estimand}
\usage{
specify_estimand(..., constraints = cc(), name = NA)
}
\arguments{
\item{...}{(character | rmlx_constraint | rmlx_constraints) \cr
Either an \code{mlguide_constraints} object or a list of \code{mlguide_constraint}s objects.
Alternatively, a character (vector) can be specified instead, which is passed to
\code{\link{parse_constraints}}.}

\item{constraints}{(rmlx_constraints) \cr a \code{mlguide_constraints} object (created with \code{\link{cc}}),
will be combined with \code{...} arguments (via\code{\link{cc}}).}

\item{name}{(character) \cr The label/name of the estimand (optional, default: NA).}
}
\value{
rmlx_estimand
}
\description{
define an estimand by a set of constraints.
}
\examples{
estimand <- specify_estimand(cc(
  "(test[['year']] - train[['year']]) \%in\% 1:2 #name=time_transferability #type=context",
  "test[['country']] != train[['country']] #name=region_transferability #type=context"
))
}
