% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_splitting_rule.R
\name{specify_splitting_rule}
\alias{specify_splitting_rule}
\title{Specify a data splitting method}
\usage{
specify_splitting_rule(method, ..., default = FALSE, name = NA)
}
\arguments{
\item{method}{(character) \cr}

\item{...}{(any) \cr named arguments to function name}

\item{default}{(logical) \cr Should default arguments be used? (default: FALSE)}

\item{name}{(character) \cr name of the data splitting method, equal to \code{method} by default.}
}
\value{
an object of class \code{rmlx_splitting_rule}
}
\description{
Specify a data splitting method
}
\examples{
\dontrun{
specify_splitting_rule()
}
specify_splitting_rule("cv")
specify_splitting_rule("cv", default = TRUE)
specify_splitting_rule("cv", n_folds = 10)
}
