% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\alias{get_data_splits}
\alias{get_data_split}
\alias{get_train_sets}
\alias{get_train_set}
\alias{get_test_sets}
\alias{get_test_set}
\title{Split data}
\usage{
get_data_splits(data, splits, idx = get_valid_idx(splits))

get_data_split(data, splits, idx)

get_train_sets(data, splits, idx = get_valid_idx(splits))

get_train_set(data, splits, idx)

get_test_sets(data, splits, idx = get_valid_idx(splits))

get_test_set(data, splits, idx)

split_data(
  data,
  splits,
  idx = get_valid_idx(splits),
  set = c("both", "train", "test"),
  simplify = length(idx) == 1
)
}
\arguments{
\item{data}{(data.frame) \cr \code{data.frame} that was used in \code{\link{derive_splits}()}}

\item{splits}{(rmlx_splits) \cr a \code{rmlx_splits} object corresponding to \code{data}}

\item{idx}{(numeric) \cr index of the target split (with length 1)}

\item{set}{(character) \cr What data sets should be retreived?}

\item{simplify}{(character) \cr Should the output be simplified?
(only possible for \code{length(idx) == 1} and default in this case)}
}
\value{
\itemize{
\item \code{get_data_splits}: retrieve list (of lists) with train and test data sets
\item \code{get_data_split}: retrieve list with single train set and test data set (\code{length(idx) == 1})
\item \code{get_train_sets}: retrieve list of train data sets
\item \code{get_train_set}: retrieve single train data set (\code{length(idx) == 1})
\item \code{get_test_sets}: retrieve list of test data sets
\item \code{get_test_set}: retrieve single test data set (\code{length(idx) == 1})
\item \code{split_data}: a list containing one or multiple train and/or test data sets (depending on \code{set} and \code{simplify})
}
}
\description{
Subset the \code{data} to retrieve the train and/or test data corresponding to \code{splits[[idx]]},
where \code{length(idx) == 1}.
}
