% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_splits.R
\name{derive_splits}
\alias{derive_splits}
\title{Derive a data split for a given estimand.}
\usage{
derive_splits(design, data)
}
\arguments{
\item{design}{(rmlx_design) \cr a \code{rmlx_design} object, i.e. a
\code{rmlx_estimand} object created via specify \code{\link{specify_estimand}} or a
\code{rmlx_splitting_rule} object created via specify \code{\link{specify_splitting_rule}}}

\item{data}{(data.frame) \cr the complete data set.}
}
\value{
(rmlx_splits)
}
\description{
Derive a data split for a given estimand.
}
